+++ diesel_cli/tests/migration_run.rs
@@ -423,6 +423,43 @@ fn migration_run_updates_schema_if_config_present() {
 }
 
 #[test]
+fn migration_run_ignores_schema_if_config_present_with_no_schema_cli_flag() {
+    let p = project("migration_run_ignores_schema_with_cli_flag")
+        .folder("migrations")
+        .file(
+            "diesel.toml",
+            r#"
+            [print_schema]
+            file = "src/my_schema.rs"
+            "#,
+        )
+        .build();
+
+    // Make sure the project is setup
+    p.command("setup").run();
+
+    p.create_migration(
+        "12345_create_users_table",
+        "CREATE TABLE users (id INTEGER PRIMARY KEY)",
+        Some("DROP TABLE users"),
+        None,
+    );
+
+    assert!(!p.has_file("src/my_schema.rs"));
+
+    let result = p.command("migration").arg("run").arg("--no-schema").run();
+
+    assert!(result.is_success(), "Result was unsuccessful {:?}", result);
+    assert!(
+        result.stdout().contains("Running migration 12345"),
+        "Unexpected stdout {}",
+        result.stdout()
+    );
+    // Schema still doesn't exist.
+    assert!(!p.has_file("src/my_schema.rs"));
+}
+
+#[test]
 fn migrations_can_be_run_with_no_config_file() {
     let p = project("migration_run_no_config_file")
         .folder("migrations")