+++ diesel_cli/tests/support/mysql_database.rs
@@ -41,7 +41,7 @@ impl Database {
 
     pub fn conn(&self) -> MysqlConnection {
         MysqlConnection::establish(&self.url)
-            .expect(&format!("Failed to open connection to {}", &self.url))
+            .expect(&format!("Failed to open connection to {}", self.url))
     }
 
     pub fn execute(&self, command: &str) {